backend/requirements.txt
# Core
fastapi
uvicorn[standard]
# Database
sqlalchemy
asyncpg
psycopg2-binary
# MongoDB
motor
pymongo
# Authentication
python-jose[cryptography]
passlib[bcrypt]
bcrypt
# Utilities
python-dotenv
python-multipart
email-validator
pytz
pydantic-settings
# HTTP
requests
httpx
aiohttp
# Task Queue
celery
redis
# Optional but helpful
python-multipart
httpx
flower==2.0.1
django-celery-beat
